backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pmc-crm-backend"
version = "0.1.0"
description = "Private Medical Center CRM backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.20,<1.0",
  "alembic>=1.14,<2.0",
  "argon2-cffi>=23.1,<26.0",
  "asyncpg>=0.29,<1.0",
  "celery[redis]>=5.4,<6.0",
  "cryptography>=42,<46",
  "email-validator>=2.2,<3.0",
  "fastapi>=0.115,<1.0",
  "greenlet>=3.1,<4.0",
  "httpx>=0.28,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "redis>=5.2,<6.0",
  "sqlalchemy>=2.0.36,<3.0",
  "uvicorn[standard]>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
  "fakeredis>=2.28,<3.0",
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<1.0",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

